refactor: Simplify ThumbnailImageLabel using coordinator pattern
- Reduced file size from 563 lines to 277 lines (51% reduction)
- Extracted complex image processing logic to existing ThumbnailCoordinator
- Removed 463 lines of complex processing, caching, and scaling code
- Delegated responsibilities to focused components:
  * Image processing → ThumbnailProcessor
  * Cache management → ThumbnailCacheManager
  * Size calculations → ThumbnailSizeCalculator
  * Event handling → ThumbnailEventHandler
- Maintained backward compatibility with existing API
- Improved maintainability through single responsibility principle
- Added backup file for reference

This refactoring follows the coordinator pattern established in the codebase,
making the thumbnail system more modular and easier to maintain.
