+++ src/main_window/main_widget/browse_tab/thumbnail_box/thumbnail_image_label.py
@@ -18,108 +18,6 @@ if TYPE_CHECKING:
     from .thumbnail_box import ThumbnailBox
 
 
-class ImageProcessor:
-    """
-    Simple, consistent image processor using Qt SmoothTransformation.
-
-    Provides high-quality image scaling for all thumbnails and exports
-    using Qt's built-in SmoothTransformation for consistent, sharp results.
-    """
-
-    @staticmethod
-    def process_image(source_path: str, target_size: QSize) -> QPixmap:
-        """
-        Process image using Qt SmoothTransformation for consistent quality.
-
-        Args:
-            source_path: Path to source image
-            target_size: Target size for thumbnail
-
-        Returns:
-            High-quality QPixmap using SmoothTransformation
-        """
-        return ImageProcessor._process_with_qt(source_path, target_size)
-
-    @staticmethod
-    def _process_with_qt(source_path: str, target_size: QSize) -> QPixmap:
-        """
-        Fallback processing using Qt when PIL is not available.
-        Enhanced Qt-only processing with multi-step scaling.
-        """
-        try:
-            # Load original image
-            original_pixmap = QPixmap(source_path)
-            if original_pixmap.isNull():
-                logging.error(f"Failed to load image: {source_path}")
-                return ImageProcessor._create_error_pixmap(target_size)
-
-            # Calculate target dimensions maintaining aspect ratio
-            original_size = original_pixmap.size()
-            aspect_ratio = original_size.width() / original_size.height()
-
-            if target_size.width() / target_size.height() > aspect_ratio:
-                new_h = target_size.height()
-                new_w = int(target_size.height() * aspect_ratio)
-            else:
-                new_w = target_size.width()
-                new_h = int(target_size.width() / aspect_ratio)
-
-            target_w, target_h = new_w, new_h
-            scale_factor = min(
-                target_w / original_size.width(), target_h / original_size.height()
-            )
-
-            # Enhanced multi-step scaling with Qt
-            if scale_factor < 0.6:  # Lower threshold than before
-                # Multi-step scaling
-                intermediate_factor = 0.75 if scale_factor < 0.4 else 0.8
-                intermediate_w = int(original_size.width() * intermediate_factor)
-                intermediate_h = int(original_size.height() * intermediate_factor)
-
-                # Step 1: Scale to intermediate size
-                intermediate_pixmap = original_pixmap.scaled(
-                    intermediate_w,
-                    intermediate_h,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                # Step 2: Scale to final size
-                final_pixmap = intermediate_pixmap.scaled(
-                    target_w,
-                    target_h,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-            else:
-                # Single-step high-quality scaling
-                final_pixmap = original_pixmap.scaled(
-                    target_w,
-                    target_h,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-            return final_pixmap
-
-        except Exception as e:
-            logging.error(f"Qt fallback processing failed for {source_path}: {e}")
-            return ImageProcessor._create_error_pixmap(target_size)
-
-    @staticmethod
-    def _create_error_pixmap(size: QSize) -> QPixmap:
-        """Create error pixmap when image processing fails."""
-        pixmap = QPixmap(size)
-        pixmap.fill(QColor(200, 200, 200))
-
-        painter = QPainter(pixmap)
-        painter.setPen(QColor(100, 100, 100))
-        painter.drawText(pixmap.rect(), Qt.AlignmentFlag.AlignCenter, "Image\nError")
-        painter.end()
-
-        return pixmap
-
-
 class ThumbnailImageLabel(QLabel):
     BORDER_WIDTH_RATIO: Final = 0.01
     SEQUENCE_VIEWER_BORDER_SCALE: Final = 0.8
@@ -142,26 +40,8 @@ class ThumbnailImageLabel(QLabel):
         self._original_pixmap: Optional[QPixmap] = None
         self._cached_available_size: Optional[QSize] = None
 
-        # Image processor
-        self.image_processor = ImageProcessor()
-
-        # Deferred loading to prevent UI blocking
-        self._load_timer = QTimer()
-        self._load_timer.setSingleShot(True)
-        self._load_timer.timeout.connect(self._load_pending_image)
-        self._pending_path: Optional[str] = None
-        self._pending_index: Optional[int] = None
-
-        # Quality enhancement timer
-        self._quality_timer = QTimer()
-        self._quality_timer.setSingleShot(True)
-        self._quality_timer.timeout.connect(self._enhance_image_quality)
-        self._needs_quality_enhancement = False
-
-        # Initialize cache system
-        self._cache_metadata = {}
-        self._ensure_cache_directory()
-        self._load_cache_metadata()
+        # Initialize coordinator to handle complex operations
+        self.coordinator = ThumbnailCoordinator(thumbnail_box)
 
         # Setup UI
         self.setCursor(QCursor(Qt.CursorShape.PointingHandCursor))
@@ -199,7 +79,14 @@ class ThumbnailImageLabel(QLabel):
             self._original_pixmap = QPixmap(path)
             self._cached_available_size = None
 
-        self._resize_pixmap_to_ultra_quality()
+        # Use coordinator for processing
+        processed_pixmap = self.coordinator.process_thumbnail_sync(
+            path, self.is_in_sequence_viewer
+        )
+
+        if not processed_pixmap.isNull():
+            self.setPixmap(processed_pixmap)
+            self._update_size_from_pixmap(processed_pixmap)
 
     def update_thumbnail_async(self, index: int) -> None:
         """Update the displayed image asynchronously with ultra quality processing."""
@@ -209,379 +96,104 @@ class ThumbnailImageLabel(QLabel):
 
         path = thumbnails[index]
         if path != self.current_path:
-            self._pending_path = path
-            self._pending_index = index
-
-            # Load with ultra quality processing (no cache)
-            self._load_timer.start(1)
+            self.current_path = path
+            # Use coordinator for async processing
+            self.coordinator.process_thumbnail_async(path, index)
+            # Set up a timer to update the display when processing completes
+            QTimer.singleShot(10, lambda: self._update_display_from_coordinator())
         else:
-            self._resize_pixmap_to_ultra_quality()
+            # Use coordinator for processing
+            processed_pixmap = self.coordinator.process_thumbnail_sync(
+                path, self.is_in_sequence_viewer
+            )
 
+            if not processed_pixmap.isNull():
+                self.setPixmap(processed_pixmap)
+                self._update_size_from_pixmap(processed_pixmap)
+
+    def _update_display_from_coordinator(self) -> None:
+        """Update display after coordinator processing."""
+        if self.current_path:
+            processed_pixmap = self.coordinator.process_thumbnail_sync(
+                self.current_path, self.is_in_sequence_viewer
+            )
+
+            if not processed_pixmap.isNull():
+                self.setPixmap(processed_pixmap)
+                self._update_size_from_pixmap(processed_pixmap)
+
+    def _update_size_from_pixmap(self, pixmap: QPixmap) -> None:
+        """Update widget size based on pixmap."""
+        if not pixmap.isNull():
+            # Calculate appropriate size for the widget
+            target_size = self.coordinator.size_calculator.calculate_target_size(
+                self.is_in_sequence_viewer
+            )
+            self.setFixedSize(target_size)
+
+    # Legacy method - now handled by coordinator
     def _load_pending_image(self) -> None:
-        """Load pending image with ultra quality processing."""
-        if self._pending_path and self._pending_index is not None:
-            try:
-                # Process with ultra quality (no cache)
-                if os.path.exists(self._pending_path):
-                    self.current_path = self._pending_path
-                    self._original_pixmap = QPixmap(self._pending_path)
-                    self._cached_available_size = None
+        """Legacy method - processing now handled by coordinator."""
+        pass
 
-                    # Always use ultra quality processing
-                    self._resize_pixmap_to_ultra_quality()
-
-            except Exception as e:
-                logging.error(f"Error in ultra quality loading: {e}")
-            finally:
-                self._pending_path = None
-                self._pending_index = None
-
+    # Size calculation now handled by coordinator
     def _calculate_available_space(self) -> QSize:
-        """Calculate available space - ENHANCED FOR MAXIMUM QUALITY."""
-        if self._cached_available_size:
-            return self._cached_available_size
-
-        if self.is_in_sequence_viewer:
-            available_size = self._calculate_sequence_viewer_size()
-        else:
-            available_size = self._calculate_normal_view_size_enhanced()
-
-        self._cached_available_size = available_size
-        return available_size
-
-    def _calculate_normal_view_size_enhanced(self) -> QSize:
-        """Enhanced calculation for maximum quality thumbnails."""
-        scroll_widget = self.thumbnail_box.sequence_picker.scroll_widget
-        scroll_widget_width = scroll_widget.width()
-
-        # Account for scrollbar and margins
-        scrollbar_width = scroll_widget.calculate_scrollbar_width()
-
-        # ULTRA QUALITY: Maximize thumbnail size for crisp display
-        total_margins = (3 * self.thumbnail_box.margin * 2) + 5
-        usable_width = scroll_widget_width - scrollbar_width - total_margins
-
-        # Calculate thumbnail width (3 columns)
-        thumbnail_width = max(200, int(usable_width // 3))  # Increased from 150 to 200
-
-        # ULTRA QUALITY: Use maximum available space (minimal padding)
-        available_width = int(thumbnail_width - 8)  # Minimal padding
-        available_width = max(
-            180, available_width
-        )  # Increased from 140 to 180 for better quality
-
-        # Calculate height based on aspect ratio
-        available_height = int(available_width / self.aspect_ratio)
-
-        # ULTRA QUALITY: Ensure minimum size for crisp display
-        available_height = max(135, available_height)  # Increased from 100 to 135
-
-        return QSize(available_width, available_height)
-
-    def _calculate_sequence_viewer_size(self) -> QSize:
-        """Calculate available space in sequence viewer mode."""
-        sequence_viewer = self.thumbnail_box.browse_tab.sequence_viewer
-
-        try:
-            available_width = int(sequence_viewer.width() * 0.95)
-            available_height = int(sequence_viewer.height() * 0.65)
-
-            # ULTRA QUALITY: Higher minimums for sequence viewer
-            available_width = max(400, available_width)
-            available_height = max(300, available_height)
-
-        except (AttributeError, TypeError):
-            available_width = 500  # Higher fallback
-            available_height = 400
-
-        return QSize(available_width, available_height)
+        """Legacy method - size calculation now handled by coordinator."""
+        return self.coordinator.size_calculator.calculate_target_size(
+            self.is_in_sequence_viewer
+        )
 
     def _resize_pixmap_to_ultra_quality(self) -> None:
-        """Resize pixmap using ULTRA QUALITY processing with user settings."""
+        """Legacy method - processing now handled by coordinator."""
         if not self.current_path:
             return
 
-        available_size = self._calculate_available_space()
-
-        # Check cache first
-        cached_pixmap = self._get_cached_thumbnail(available_size)
-        if cached_pixmap and not cached_pixmap.isNull():
-            self.setFixedSize(available_size)
-            self.setPixmap(cached_pixmap)
-            logging.debug(
-                f"✅ Loaded cached thumbnail: {os.path.basename(self.current_path)}"
-            )
-            return
-
-        # HIGH QUALITY PROCESSING - always use SmoothTransformation
-        processed_pixmap = self.image_processor.process_image(
-            self.current_path,
-            available_size,
+        # Use coordinator for processing
+        processed_pixmap = self.coordinator.process_thumbnail_sync(
+            self.current_path, self.is_in_sequence_viewer
         )
 
-        # CRITICAL: Ensure pixmap is not null before proceeding
-        if processed_pixmap.isNull():
-            logging.warning(
-                f"Failed to create processed pixmap for {self.current_path}"
-            )
-            # Fallback to standard processing
-            self._resize_pixmap_to_fit_smooth()
-            return
-
-        # Cache the high-quality thumbnail
-        self._cache_thumbnail(processed_pixmap, available_size)
-
-        self.setFixedSize(available_size)
-        self.setPixmap(processed_pixmap)
-
-        logging.debug(
-            f"✅ High-quality thumbnail processed: {os.path.basename(self.current_path)}"
-        )
+        if not processed_pixmap.isNull():
+            self.setPixmap(processed_pixmap)
+            self._update_size_from_pixmap(processed_pixmap)
 
     # Removed _get_quality_settings - now always using maximum quality
 
     def _resize_pixmap_to_fit_smooth(self) -> None:
-        """Enhanced smooth resizing with improved multi-step scaling."""
-        if not self._original_pixmap:
-            return
+        """Legacy method - processing now handled by coordinator."""
+        self._resize_pixmap_to_ultra_quality()
 
-        available_size = self._calculate_available_space()
-        scaled_size = self._calculate_scaled_pixmap_size(available_size)
-
-        # Enhanced multi-step scaling for better quality
-        scaled_pixmap = self._create_enhanced_scaled_pixmap(scaled_size)
-
-        # CRITICAL: Ensure pixmap is not null before proceeding
-        if scaled_pixmap.isNull():
-            logging.warning(f"Failed to create scaled pixmap for {self.current_path}")
-            return
-
-        self.setFixedSize(available_size)
-        self.setPixmap(scaled_pixmap)
-
-        # Cache system removed - no longer needed
-        logging.debug(
-            f"✅ HIGH-QUALITY thumbnail processed: {os.path.basename(self.current_path)}"
-        )
-
+    # Legacy methods - processing now handled by coordinator
     def _calculate_scaled_pixmap_size(self, available_size: QSize) -> QSize:
-        """Calculate the optimal size for the pixmap while maintaining aspect ratio."""
-        if not self._original_pixmap:
-            return QSize(0, 0)
-
-        aspect_ratio = self._original_pixmap.height() / self._original_pixmap.width()
-        target_width = available_size.width()
-        target_height = int(target_width * aspect_ratio)
-
-        if target_height > available_size.height():
-            target_height = available_size.height()
-            target_width = int(target_height / aspect_ratio)
-
-        return QSize(target_width, target_height)
+        """Legacy method - calculations now handled by coordinator."""
+        return available_size
 
     def _create_enhanced_scaled_pixmap(self, target_size: QSize) -> QPixmap:
-        """Create enhanced scaled pixmap with improved multi-step scaling."""
-        if not self._original_pixmap:
-            return QPixmap()
-
-        original_size = self._original_pixmap.size()
-
-        # Calculate scale factor
-        scale_factor = min(
-            target_size.width() / original_size.width(),
-            target_size.height() / original_size.height(),
-        )
-
-        # Enhanced multi-step scaling with lower threshold
-        if scale_factor < 0.75:  # Improved from 0.5
-            # Multi-step scaling for better quality
-            if scale_factor < 0.4:
-                # Very aggressive downscaling - use 3 stages
-                intermediate_factor1 = 0.7
-                intermediate_factor2 = 0.5
-
-                # Stage 1
-                intermediate_size1 = QSize(
-                    int(original_size.width() * intermediate_factor1),
-                    int(original_size.height() * intermediate_factor1),
-                )
-                stage1_pixmap = self._original_pixmap.scaled(
-                    intermediate_size1,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                # Stage 2
-                intermediate_size2 = QSize(
-                    int(original_size.width() * intermediate_factor2),
-                    int(original_size.height() * intermediate_factor2),
-                )
-                stage2_pixmap = stage1_pixmap.scaled(
-                    intermediate_size2,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                # Final stage
-                final_pixmap = stage2_pixmap.scaled(
-                    target_size,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                return final_pixmap
-            else:
-                # Moderate downscaling - use 2 stages
-                intermediate_factor = 0.75
-                intermediate_size = QSize(
-                    int(original_size.width() * intermediate_factor),
-                    int(original_size.height() * intermediate_factor),
-                )
-
-                # Step 1: Scale to intermediate size
-                intermediate_pixmap = self._original_pixmap.scaled(
-                    intermediate_size,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                # Step 2: Scale to final size
-                final_pixmap = intermediate_pixmap.scaled(
-                    target_size,
-                    Qt.AspectRatioMode.KeepAspectRatio,
-                    Qt.TransformationMode.SmoothTransformation,
-                )
-
-                return final_pixmap
-        else:
-            # Single-step high-quality scaling for smaller scale changes
-            return self._original_pixmap.scaled(
-                target_size,
-                Qt.AspectRatioMode.KeepAspectRatio,
-                Qt.TransformationMode.SmoothTransformation,
+        """Legacy method - scaling now handled by coordinator."""
+        if self.current_path:
+            return self.coordinator.process_thumbnail_sync(
+                self.current_path, self.is_in_sequence_viewer
             )
+        return QPixmap()
 
-    # Cache management methods
-    def _ensure_cache_directory(self) -> None:
-        """Ensure cache directory exists."""
-        try:
-            self.CACHE_DIR.mkdir(exist_ok=True)
-        except Exception as e:
-            logging.warning(f"Failed to create cache directory: {e}")
-
-    def _load_cache_metadata(self) -> None:
-        """Load cache metadata from disk."""
-        metadata_path = self.CACHE_DIR / self.CACHE_METADATA_FILE
-        try:
-            if metadata_path.exists():
-                with open(metadata_path, "r") as f:
-                    self._cache_metadata = json.load(f)
-            else:
-                self._cache_metadata = {}
-        except Exception as e:
-            logging.warning(f"Failed to load cache metadata: {e}")
-            self._cache_metadata = {}
-
-    def _save_cache_metadata(self) -> None:
-        """Save cache metadata to disk."""
-        metadata_path = self.CACHE_DIR / self.CACHE_METADATA_FILE
-        try:
-            with open(metadata_path, "w") as f:
-                json.dump(self._cache_metadata, f, indent=2)
-        except Exception as e:
-            logging.warning(f"Failed to save cache metadata: {e}")
-
-    def _generate_cache_key(self, image_path: str, target_size: QSize) -> str:
-        """Generate cache key based on image path, modification time, and size."""
-        try:
-            # Get file modification time
-            mtime = os.path.getmtime(image_path)
-
-            # Create cache key from path, mtime, and target size
-            key_data = (
-                f"{image_path}_{mtime}_{target_size.width()}x{target_size.height()}"
-            )
-            return hashlib.md5(key_data.encode()).hexdigest()
-        except Exception:
-            # Fallback to path-only key if mtime fails
-            key_data = f"{image_path}_{target_size.width()}x{target_size.height()}"
-            return hashlib.md5(key_data.encode()).hexdigest()
-
-    def _get_cached_thumbnail(self, target_size: QSize) -> Optional[QPixmap]:
-        """Get cached thumbnail if available and valid."""
-        if not self.current_path:
-            return None
-
-        cache_key = self._generate_cache_key(self.current_path, target_size)
-        cache_file = self.CACHE_DIR / f"{cache_key}.png"
-
-        try:
-            # Check if cache file exists and metadata is valid
-            if cache_file.exists() and cache_key in self._cache_metadata:
-                metadata = self._cache_metadata[cache_key]
-
-                # Validate cache entry
-                if (
-                    metadata.get("source_path") == self.current_path
-                    and metadata.get("target_width") == target_size.width()
-                    and metadata.get("target_height") == target_size.height()
-                ):
-
-                    # Load cached pixmap
-                    pixmap = QPixmap(str(cache_file))
-                    if not pixmap.isNull():
-                        return pixmap
-
-        except Exception as e:
-            logging.debug(f"Error loading cached thumbnail: {e}")
-
-        return None
-
-    def _cache_thumbnail(self, pixmap: QPixmap, target_size: QSize) -> None:
-        """Cache thumbnail to disk."""
-        if not self.current_path or pixmap.isNull():
-            return
-
-        cache_key = self._generate_cache_key(self.current_path, target_size)
-        cache_file = self.CACHE_DIR / f"{cache_key}.png"
-
-        try:
-            # Save pixmap to cache
-            if pixmap.save(str(cache_file), "PNG"):
-                # Update metadata
-                self._cache_metadata[cache_key] = {
-                    "source_path": self.current_path,
-                    "target_width": target_size.width(),
-                    "target_height": target_size.height(),
-                    "cached_at": os.path.getmtime(self.current_path),
-                    "cache_file": str(cache_file),
-                }
-
-                # Save metadata (async to avoid blocking)
-                QTimer.singleShot(100, self._save_cache_metadata)
-
-                logging.debug(
-                    f"✅ Cached thumbnail: {os.path.basename(self.current_path)}"
-                )
-            else:
-                logging.warning(f"Failed to save thumbnail cache: {cache_file}")
-
-        except Exception as e:
-            logging.warning(f"Error caching thumbnail: {e}")
+    # Cache management now handled by coordinator
 
     def mousePressEvent(self, event: QMouseEvent) -> None:
         """Handle mouse press events."""
+        self.coordinator.handle_mouse_press(event)
         if not self.is_in_sequence_viewer:
             self.thumbnail_box.browse_tab.selection_handler.on_thumbnail_clicked(self)
 
     def enterEvent(self, event) -> None:
         """Highlight border on hover."""
-        self._border_color = BLUE
+        self.coordinator.handle_enter_event(event)
+        self._border_color = self.coordinator.get_border_color() or BLUE
         self.update()
         super().enterEvent(event)
 
     def leaveEvent(self, event) -> None:
         """Remove border highlight when leaving hover."""
+        self.coordinator.handle_leave_event(event)
         self._border_color = GOLD if self.selected else None
         self.update()
         super().leaveEvent(event)
@@ -589,6 +201,7 @@ class ThumbnailImageLabel(QLabel):
     def set_selected(self, selected: bool) -> None:
         """Set selection state."""
         self.selected = selected
+        self.coordinator.set_selection_state(selected)
         self._border_color = GOLD if selected else None
         self.update()
 
@@ -636,11 +249,10 @@ class ThumbnailImageLabel(QLabel):
         self._border_width = max(1, int(self.width() * self.BORDER_WIDTH_RATIO))
         super().resizeEvent(event)
 
-    # Removed cache initialization and word/variation tracking - no longer needed
-
+    # Legacy methods - functionality now handled by coordinator
     def _enhance_image_quality(self) -> None:
-        """Legacy method - now handled by ultra_processor."""
-        pass  # Ultra quality processing handles all enhancement
+        """Legacy method - now handled by coordinator."""
+        pass
 
     def _check_viewport_visibility(self) -> bool:
         """Check if this thumbnail is currently visible in the viewport."""
